Remove MovieList and Restructure page
